+++ CHANGELOG.md
@@ -65,6 +65,7 @@ The types of changes are:
 * DRP exerise endpoint [#496](https://github.com/ethyca/fidesops/pull/496)
 * Frontend for privacy request denial reaons [#480](https://github.com/ethyca/fidesops/pull/480)
 * Publish Fidesops to Pypi [#491](https://github.com/ethyca/fidesops/pull/491)
+* DRP data rights endpoint [#526](https://github.com/ethyca/fidesops/pull/526)
 
 
 ### Changed